+++ src/main/java/org/saucedemo/testresults/TestResult.java
@@ -6,7 +6,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.openqa.selenium.JavascriptExecutor;
 import org.saucedemo.factories.EnvConfigFactory;
 
-import static org.saucedemo.hosts.localhost.android.AvailableDevices.freeDevice;
+import static org.saucedemo.hosts.localhost.android.EmulatorDevicePicker.freeDevice;
 import static org.saucedemo.testresults.RunnerExtension.getClassName;
 import static org.saucedemo.testresults.RunnerExtension.getReason;
 import static org.saucedemo.testresults.RunnerExtension.getTestName;
